Modularizado el rating para poder llamarlo tanto en la tarjeta como en el detalle
